requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2020.12.5
dj-database-url==0.5.0
Django==3.2.3
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
persian-tools==0.0.9
psycopg2-binary==2.8.6
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
